+++ packages/cli/src/commands/bisect/index.ts
@@ -2,15 +2,15 @@ import open from 'open';
 import execa from 'execa';
 import plural from 'pluralize';
 import { resolve } from 'path';
-import chalk, { Chalk } from 'chalk';
+import chalk, { type Chalk } from 'chalk';
 import { URLSearchParams, parse } from 'url';
 
 import box from '../../util/output/box';
 import formatDate from '../../util/format-date';
 import link from '../../util/output/link';
 import { parseArguments } from '../../util/get-args';
-import Client from '../../util/client';
-import { Deployment } from '@vercel-internals/types';
+import type Client from '../../util/client';
+import type { Deployment } from '@vercel-internals/types';
 import { normalizeURL } from '../../util/bisect/normalize-url';
 import getScope from '../../util/get-scope';
 import getDeployment from '../../util/get-deployment';
@@ -18,15 +18,13 @@ import { help } from '../help';
 import { bisectCommand } from './command';
 import { getFlagsSpecification } from '../../util/get-flags-specification';
 import handleError from '../../util/handle-error';
+import output from '../../output-manager';
+import { BisectTelemetryClient } from '../../util/telemetry/commands/bisect';
 
 interface Deployments {
   deployments: Deployment[];
 }
 export default async function bisect(client: Client): Promise<number> {
-  const { output } = client;
-  const scope = await getScope(client);
-  const { contextName } = scope;
-
   let parsedArgs = null;
 
   const flagsSpecification = getFlagsSpecification(bisectCommand.options);
@@ -38,10 +36,26 @@ export default async function bisect(client: Client): Promise<number> {
     return 1;
   }
 
+  const telemetry = new BisectTelemetryClient({
+    opts: {
+      store: client.telemetryEventStore,
+    },
+  });
+
   if (parsedArgs.flags['--help']) {
+    telemetry.trackCliFlagHelp('bisect');
     output.print(help(bisectCommand, { columns: client.stderr.columns }));
     return 2;
   }
+
+  telemetry.trackCliOptionGood(parsedArgs.flags['--good']);
+  telemetry.trackCliOptionBad(parsedArgs.flags['--bad']);
+  telemetry.trackCliOptionPath(parsedArgs.flags['--path']);
+  telemetry.trackCliOptionRun(parsedArgs.flags['--run']);
+  telemetry.trackCliFlagOpen(parsedArgs.flags['--open']);
+
+  const scope = await getScope(client);
+  const { contextName } = scope;
 
   let bad =
     parsedArgs.flags['--bad'] ||
@@ -306,7 +320,7 @@ export default async function bisect(client: Client): Promise<number> {
 
   output.print('\n');
 
-  let result = [
+  const result = [
     chalk.bold(
       `The first bad deployment is: ${link(`https://${lastBad.url}`)}`
     ),